refactor: GameController에서 한 게임이 진행될 수 있도록, 한 이닝이 진행될 수 있도록 변경
