+++ src/main/kotlin/baseball/Application.kt
@@ -1,4 +1,4 @@
-package racingcar
+package baseball
 
 import baseball.domain.GameController
 import baseball.view.InputView
@@ -7,5 +7,4 @@ import baseball.view.OutputView
 fun main() {
     val controller = GameController(InputView(), OutputView())
     controller.start()
-    controller.run()
 }
